Develop

+++ Vendas.Infra/Mappings/UsuarioMap.cs
@@ -14,10 +14,10 @@ namespace Vendas.Infra.Mappings
             builder.HasIndex(x => x.Login).IsUnique();
             
 
-            builder.Property(x => x.IdUsuario).HasColumnName("idUsuario").HasColumnType("bigint").IsRequired();
+            builder.Property(x => x.IdUsuario).HasColumnName("idUsuario").HasColumnType("int").IsRequired();
             builder.Property(x => x.NomeUsuario).HasColumnName("nomeUsuario").HasColumnType("varchar(60)").IsRequired();
             builder.Property(x => x.Senha).HasColumnName("senha").HasColumnType("varchar(60)").IsRequired();
-            builder.Property(x => x.EhAdm).HasColumnName("ehAdm").HasColumnType("smallint").IsRequired();
+            builder.Property(x => x.EhAdm).HasColumnName("ehAdm").HasColumnType("int").IsRequired();
             builder.Property(x => x.Login).HasColumnName("login").HasColumnType("varchar(60)").IsRequired();
         }
     }
